Add files via upload
In this data set, I analyzed a data set of a shipping company who ships across the globe. I calculated the total revenue, number of customers, number of orders, count of products, count of categories of products, and the number of countries they ship to. I visualized the total sales by month and quarter using a line graph to show the time series, the top 10 customers using the sum of revenues using a bar chart ,the top 5 products using revenue, top 5 products by quantity sold etc. I added some slicers to make it an interactive dashboard and I used a text box to add the headings and cards to display the important numbers.